Re-arrange component_wrapper files and jest coverage actualization (#17353)
