+++ js/renovation/component_wrapper/__tests__/data_grid.test.tsx
@@ -4,7 +4,7 @@ import DataGridComponent from '../data_grid';
 const mockBeginUpdate = jest.fn();
 const mockEndUpdate = jest.fn();
 
-jest.mock('../component', () => class {
+jest.mock('../common/component', () => class {
   beginUpdate() {
     mockBeginUpdate();
   }
