+++ js/renovation/component_wrapper/common/__tests__/utils/test_components/base_test_widget.ts
