+++ js/renovation/component_wrapper/common/__tests__/utils/test_components/empty_test_widget.tsx
@@ -4,6 +4,8 @@ import {
   JSXComponent,
 } from '@devextreme-generator/declarations';
 
+// import * as Declarations from '@devextreme-generator/declarations';
+
 export const view = (): JSX.Element => <div />;
 
 @ComponentBindings()
