+++ js/renovation/component_wrapper/common/__tests__/utils/test_components/options_check_widget.tsx
