+++ js/renovation/component_wrapper/common/__tests__/utils/test_components/templated_test_widget.tsx
