+++ js/renovation/component_wrapper/common/component.ts
@@ -1,16 +1,16 @@
 /* eslint-disable */
 import { render, createRef, RefObject } from "inferno";
 import { createElement } from 'inferno-create-element';
-import $ from '../../core/renderer';
-import domAdapter from '../../core/dom_adapter';
-import DOMComponent from '../../core/dom_component';
-import { extend } from '../../core/utils/extend';
-import { getPublicElement } from '../../core/element';
-import { isDefined, isRenderer } from '../../core/utils/type';
+import $ from '../../../core/renderer';
+import domAdapter from '../../../core/dom_adapter';
+import DOMComponent from '../../../core/dom_component';
+import { extend } from '../../../core/utils/extend';
+import { getPublicElement } from '../../../core/element';
+import { isDefined, isRenderer } from '../../../core/utils/type';
 
 import { InfernoEffectHost, hydrate } from "@devextreme/vdom";
 import { TemplateWrapper } from "./template_wrapper";
-import { updatePropsImmutable } from './utils';
+import { updatePropsImmutable } from '../utils/update-props-immutable';
 
 
 const setDefaultOptionValue = (options, defaultValueGetter) => (name) => {
