+++ js/renovation/component_wrapper/common/template_wrapper.ts
@@ -1,13 +1,13 @@
 import { InfernoComponent, InfernoEffect } from '@devextreme/vdom';
 // eslint-disable-next-line spellcheck/spell-checker
 import { findDOMfromVNode } from 'inferno';
-import $ from '../../core/renderer';
-import domAdapter from '../../core/dom_adapter';
-import { getPublicElement } from '../../core/element';
-import { removeDifferentElements } from './utils';
-import Number from '../../core/polyfills/number';
-import { FunctionTemplate } from '../../core/templates/function_template';
-import { EffectReturn } from '../utils/effect_return';
+import $ from '../../../core/renderer';
+import domAdapter from '../../../core/dom_adapter';
+import { getPublicElement } from '../../../core/element';
+import { removeDifferentElements } from '../utils/utils';
+import Number from '../../../core/polyfills/number';
+import { FunctionTemplate } from '../../../core/templates/function_template';
+import { EffectReturn } from '../../utils/effect_return';
 
 interface TemplateWrapperProps {
   template: FunctionTemplate;
