+++ js/renovation/component_wrapper/data_grid.ts
@@ -1,8 +1,8 @@
 /* eslint-disable */
-import Component from './component';
+import Component from './common/component';
 import type { DataGrid } from '../ui/grids/data_grid/data_grid';
 import gridCore from '../../ui/data_grid/ui.data_grid.core';
-import { updatePropsImmutable } from './utils';
+import { updatePropsImmutable } from "./utils/update-props-immutable";
 
 export default class DataGridWrapper extends Component {
     _onInitialized!: Function;
