+++ js/renovation/component_wrapper/utils/__tests__/update-props-immutable.test.ts
@@ -1,4 +1,4 @@
-import { updatePropsImmutable } from '../utils';
+import { updatePropsImmutable } from '../update-props-immutable';
 
 class Dummy {
   nestedProp: string;
@@ -52,6 +52,16 @@ it('change array option', () => {
   const option = { dataSource };
   const props = { dataSource: option.dataSource };
   updatePropsImmutable(props, option, 'dataSource', 'dataSource');
+
+  expect(props.dataSource).toBe(dataSource);
+});
+
+it('wrong array index', () => {
+  const dataSource = [];
+
+  const option = { dataSource };
+  const props = { dataSource: option.dataSource };
+  updatePropsImmutable(props, option, 'dataSource', 'dataSource["a"]');
 
   expect(props.dataSource).toBe(dataSource);
 });
